move heart rate data to file handle
