+++ pom.xml
@@ -5,7 +5,7 @@
 
     <groupId>org.sagebionetworks</groupId>
     <artifactId>Bridge-FitBit-Worker</artifactId>
-    <version>1.0.3</version>
+    <version>1.0.4</version>
 
     <properties>
         <aws.version>1.10.56</aws.version>
@@ -40,7 +40,7 @@
         <dependency>
             <groupId>org.sagebionetworks</groupId>
             <artifactId>bridge-base</artifactId>
-            <version>2.7.8</version>
+            <version>2.7.9</version>
         </dependency>
         <dependency>
             <groupId>org.sagebionetworks.bridge</groupId>
